backend/requirements.txt
pyramid==2.0.2
pyramid_jwt==1.6.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
wsgicors==0.7.0
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
bcrypt==4.1.2
waitress==2.1.2
transaction==4.0
zope.sqlalchemy==3.1
pyramid_tm==2.5
pytest-alembic==0.10.7
webtest==3.0.0
pyramid_openapi3==0.20.1
hupper==1.12
